@@ -4,8 +4,10 @@
   "description": "A musical notation helper",
   "main": "main.js",
   "scripts": {
-    "start": "electron .",
-    "rebuild": "electron-rebuild -f -w midi"
+    "build": "tsc && copyfiles -u 1 \"public/**/*\" dist",
+    "watch": "tsc -w",
+    "lint": "eslint -c .eslintrc --ext .ts ./src",
+    "start": "npm run build && electron ./dist/main.js"
   },
   "keywords": [
     "piano",
@@ -17,19 +19,26 @@
   "author": "B. Jordan",
   "license": "MIT",
   "devDependencies": {
-    "electron": "^4.0.1",
-    "electron-reloader": "^0.3.0"
+    "@types/css-font-loading-module": "^0.0.8",
+    "@types/jquery": "^3.5.16",
+    "@types/vexflow": "^1.2.38",
+    "@typescript-eslint/eslint-plugin": "^4.27.0",
+    "@typescript-eslint/parser": "^4.27.0",
+    "copyfiles": "^2.4.1",
+    "electron": "^13.1.6",
+    "electron-reload": "^2.0.0-alpha.1",
+    "eslint": "^7.28.0",
+    "typescript": "^4.3.4"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/blairjordan/pianoboi.git"
   },
   "dependencies": {
-    "electron-rebuild": "^1.8.2",
-    "jquery": "^3.3.1",
-    "jzz": "^0.8.0",
+    "@tonaljs/tonal": "^4.10.0",
+    "jquery": "^3.7.0",
     "tonal": "^2.1.2",
-    "update-electron-app": "^1.3.0",
-    "vexflow": "^1.2.84"
+    "vexflow": "^4.1.0",
+    "webmidi": "^3.1.6"
   }
 }
